ACRE spectate toggle button
