+++ addons/zeusMonitor/XEH_PREP.hpp
@@ -1,3 +1,4 @@
 TRACE_1("",QUOTE(ADDON));
 
+PREP(toggleAcreSepcator);
 PREP(zeusInterfaceOpenedEH);
