+++ addons/zeusMonitor/functions/fnc_toggleAcreSepcator.sqf
@@ -1,0 +1,5 @@
+#include "script_component.hpp"
+
+TRACE_1("params",_this);
+
+[!ACRE_IS_SPECTATOR] call acre_api_fnc_setSpectator;
